await async tracking callbacks when navigating
