+++ packages/sdk-js/src/feature-repository.ts
@@ -6,7 +6,7 @@ import {
   Helpers,
   Polyfills,
 } from "./types/growthbook";
-import { getPolyfills } from "./util";
+import { getPolyfills, promiseTimeout } from "./util";
 import type { GrowthBook } from ".";
 
 type CacheEntry = {
@@ -272,31 +272,6 @@ function getCacheKey(instance: GrowthBook): string {
     fv,
     url,
   })}`;
-}
-
-// Guarantee the promise always resolves within {timeout} ms
-// Resolved value will be `null` when there's an error or it takes too long
-// Note: The promise will continue running in the background, even if the timeout is hit
-function promiseTimeout<T>(
-  promise: Promise<T>,
-  timeout?: number
-): Promise<T | null> {
-  return new Promise((resolve) => {
-    let resolved = false;
-    let timer: NodeJS.Timeout | undefined;
-    const finish = (data?: T) => {
-      if (resolved) return;
-      resolved = true;
-      timer && clearTimeout(timer);
-      resolve(data || null);
-    };
-
-    if (timeout) {
-      timer = setTimeout(() => finish(), timeout);
-    }
-
-    promise.then((data) => finish(data)).catch(() => finish());
-  });
 }
 
 // Populate cache from localStorage (if available)
