+++ packages/sdk-js/src/util.ts
@@ -396,3 +396,28 @@ export function getAutoExperimentChangeType(
 
   return "unknown";
 }
+
+// Guarantee the promise always resolves within {timeout} ms
+// Resolved value will be `null` when there's an error or it takes too long
+// Note: The promise will continue running in the background, even if the timeout is hit
+export async function promiseTimeout<T>(
+  promise: Promise<T>,
+  timeout?: number
+): Promise<T | null> {
+  return new Promise((resolve) => {
+    let resolved = false;
+    let timer: NodeJS.Timeout | undefined;
+    const finish = (data?: T) => {
+      if (resolved) return;
+      resolved = true;
+      timer && clearTimeout(timer);
+      resolve(data || null);
+    };
+
+    if (timeout) {
+      timer = setTimeout(() => finish(), timeout);
+    }
+
+    promise.then((data) => finish(data)).catch(() => finish());
+  });
+}
